\name{getBrainStarsMarker}
\alias{getBrainStarsMarker}
\title{Get brain region specific markers}
\usage{
  getBrainStarsMarker(keyword)
}
\arguments{
  \item{keyword}{keyword}
}
\description{
  This function get brain region specific markers in
  BrainStars
}
\examples{
my.genes.json <- getBrainStarsMarker("high/LS/count")
}
\keyword{bio}

